+++ instrumentation/opentelemetry-instrumentation-celery/src/opentelemetry/instrumentation/celery/__init__.py
@@ -52,16 +52,16 @@ API
 """
 
 import logging
-import signal
 from collections.abc import Iterable
 
 from celery import signals  # pylint: disable=no-name-in-module
 
-from opentelemetry import propagators, trace
+from opentelemetry import trace
 from opentelemetry.instrumentation.celery import utils
 from opentelemetry.instrumentation.celery.version import __version__
 from opentelemetry.instrumentation.instrumentor import BaseInstrumentor
-from opentelemetry.trace.propagation.textmap import DictGetter
+from opentelemetry.propagate import extract, inject
+from opentelemetry.propagators.textmap import DictGetter
 from opentelemetry.trace.status import Status, StatusCode
 
 logger = logging.getLogger(__name__)
@@ -128,7 +128,7 @@ class CeleryInstrumentor(BaseInstrumentor):
             return
 
         request = task.request
-        tracectx = propagators.extract(carrier_getter, request) or None
+        tracectx = extract(carrier_getter, request) or None
 
         logger.debug("prerun signal start task_id=%s", task_id)
 
@@ -137,8 +137,8 @@ class CeleryInstrumentor(BaseInstrumentor):
             operation_name, context=tracectx, kind=trace.SpanKind.CONSUMER
         )
 
-        activation = self._tracer.use_span(span, end_on_exit=True)
-        activation.__enter__()
+        activation = trace.use_span(span, end_on_exit=True)
+        activation.__enter__()  # pylint: disable=E1101
         utils.attach_span(task, task_id, (span, activation))
 
     @staticmethod
@@ -186,13 +186,14 @@ class CeleryInstrumentor(BaseInstrumentor):
             span.set_attribute(_TASK_NAME_KEY, task.name)
             utils.set_attributes_from_context(span, kwargs)
 
-        activation = self._tracer.use_span(span, end_on_exit=True)
-        activation.__enter__()
+        activation = trace.use_span(span, end_on_exit=True)
+        activation.__enter__()  # pylint: disable=E1101
+
         utils.attach_span(task, task_id, (span, activation), is_publish=True)
 
         headers = kwargs.get("headers")
         if headers:
-            propagators.inject(type(headers).__setitem__, headers)
+            inject(type(headers).__setitem__, headers)
 
     @staticmethod
     def _trace_after_publish(*args, **kwargs):
@@ -208,7 +209,7 @@ class CeleryInstrumentor(BaseInstrumentor):
             logger.warning("no existing span found for task_id=%s", task_id)
             return
 
-        activation.__exit__(None, None, None)
+        activation.__exit__(None, None, None)  # pylint: disable=E1101
         utils.detach_span(task, task_id, is_publish=True)
 
     @staticmethod