+++ instrumentation/opentelemetry-instrumentation-requests/tests/test_requests_integration.py
@@ -19,13 +19,13 @@ import httpretty
 import requests
 
 import opentelemetry.instrumentation.requests
-from opentelemetry import context, propagators, trace
+from opentelemetry import context, trace
 from opentelemetry.instrumentation.requests import RequestsInstrumentor
+from opentelemetry.propagate import get_global_textmap, set_global_textmap
 from opentelemetry.sdk import resources
-from opentelemetry.sdk.util import get_dict_as_key
 from opentelemetry.test.mock_textmap import MockTextMapPropagator
 from opentelemetry.test.test_base import TestBase
-from opentelemetry.trace.status import StatusCode
+from opentelemetry.trace import StatusCode
 
 
 class InvalidResponseObjectException(Exception):
@@ -86,31 +86,11 @@ class RequestsIntegrationTestBase(abc.ABC):
             },
         )
 
-        self.assertIs(span.status.status_code, trace.status.StatusCode.UNSET)
+        self.assertIs(span.status.status_code, trace.StatusCode.UNSET)
 
         self.check_span_instrumentation_info(
             span, opentelemetry.instrumentation.requests
         )
-
-        self.assertIsNotNone(RequestsInstrumentor().meter)
-        self.assertEqual(len(RequestsInstrumentor().meter.instruments), 1)
-        recorder = list(RequestsInstrumentor().meter.instruments.values())[0]
-        match_key = get_dict_as_key(
-            {
-                "http.flavor": "1.1",
-                "http.method": "GET",
-                "http.status_code": "200",
-                "http.url": "http://httpbin.org/status/200",
-            }
-        )
-        for key in recorder.bound_instruments.keys():
-            self.assertEqual(key, match_key)
-            # pylint: disable=protected-access
-            bound = recorder.bound_instruments.get(key)
-            for view_data in bound.view_datas:
-                self.assertEqual(view_data.labels, key)
-                self.assertEqual(view_data.aggregator.current.count, 1)
-                self.assertGreaterEqual(view_data.aggregator.current.sum, 0)
 
     def test_name_callback(self):
         def name_callback(method, url):
@@ -150,7 +130,7 @@ class RequestsIntegrationTestBase(abc.ABC):
         self.assertEqual(span.attributes.get("http.status_text"), "Not Found")
 
         self.assertIs(
-            span.status.status_code, trace.status.StatusCode.ERROR,
+            span.status.status_code, trace.StatusCode.ERROR,
         )
 
     def test_uninstrument(self):
@@ -212,9 +192,9 @@ class RequestsIntegrationTestBase(abc.ABC):
             self.assertFalse(mock_span.set_status.called)
 
     def test_distributed_context(self):
-        previous_propagator = propagators.get_global_textmap()
+        previous_propagator = get_global_textmap()
         try:
-            propagators.set_global_textmap(MockTextMapPropagator())
+            set_global_textmap(MockTextMapPropagator())
             result = self.perform_request(self.URL)
             self.assertEqual(result.text, "Hello!")
 
@@ -233,7 +213,7 @@ class RequestsIntegrationTestBase(abc.ABC):
             )
 
         finally:
-            propagators.set_global_textmap(previous_propagator)
+            set_global_textmap(previous_propagator)
 
     def test_span_callback(self):
         RequestsInstrumentor().uninstrument()
@@ -289,23 +269,6 @@ class RequestsIntegrationTestBase(abc.ABC):
         )
         self.assertEqual(span.status.status_code, StatusCode.ERROR)
 
-        self.assertIsNotNone(RequestsInstrumentor().meter)
-        self.assertEqual(len(RequestsInstrumentor().meter.instruments), 1)
-        recorder = list(RequestsInstrumentor().meter.instruments.values())[0]
-        match_key = get_dict_as_key(
-            {
-                "http.method": "GET",
-                "http.url": "http://httpbin.org/status/200",
-            }
-        )
-        for key in recorder.bound_instruments.keys():
-            self.assertEqual(key, match_key)
-            # pylint: disable=protected-access
-            bound = recorder.bound_instruments.get(key)
-            for view_data in bound.view_datas:
-                self.assertEqual(view_data.labels, key)
-                self.assertEqual(view_data.aggregator.current.count, 1)
-
     mocked_response = requests.Response()
     mocked_response.status_code = 500
     mocked_response.reason = "Internal Server Error"
@@ -323,23 +286,6 @@ class RequestsIntegrationTestBase(abc.ABC):
             span.attributes, {"http.method": "GET", "http.url": self.URL}
         )
         self.assertEqual(span.status.status_code, StatusCode.ERROR)
-
-        self.assertIsNotNone(RequestsInstrumentor().meter)
-        self.assertEqual(len(RequestsInstrumentor().meter.instruments), 1)
-        recorder = list(RequestsInstrumentor().meter.instruments.values())[0]
-        match_key = get_dict_as_key(
-            {
-                "http.method": "GET",
-                "http.url": "http://httpbin.org/status/200",
-            }
-        )
-        for key in recorder.bound_instruments.keys():
-            self.assertEqual(key, match_key)
-            # pylint: disable=protected-access
-            bound = recorder.bound_instruments.get(key)
-            for view_data in bound.view_datas:
-                self.assertEqual(view_data.labels, key)
-                self.assertEqual(view_data.aggregator.current.count, 1)
 
     mocked_response = requests.Response()
     mocked_response.status_code = 500
@@ -364,23 +310,6 @@ class RequestsIntegrationTestBase(abc.ABC):
             },
         )
         self.assertEqual(span.status.status_code, StatusCode.ERROR)
-        self.assertIsNotNone(RequestsInstrumentor().meter)
-        self.assertEqual(len(RequestsInstrumentor().meter.instruments), 1)
-        recorder = list(RequestsInstrumentor().meter.instruments.values())[0]
-        match_key = get_dict_as_key(
-            {
-                "http.method": "GET",
-                "http.status_code": "500",
-                "http.url": "http://httpbin.org/status/200",
-            }
-        )
-        for key in recorder.bound_instruments.keys():
-            self.assertEqual(key, match_key)
-            # pylint: disable=protected-access
-            bound = recorder.bound_instruments.get(key)
-            for view_data in bound.view_datas:
-                self.assertEqual(view_data.labels, key)
-                self.assertEqual(view_data.aggregator.current.count, 1)
 
     @mock.patch("requests.adapters.HTTPAdapter.send", side_effect=Exception)
     def test_requests_basic_exception(self, *_, **__):