+++ CHANGELOG.md
@@ -4,13 +4,37 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/open-telemetry/opentelemetry-python-contrib/compare/v0.17b0...HEAD)
+## [Unreleased](https://github.com/open-telemetry/opentelemetry-python-contrib/compare/v0.18b0...HEAD)
+- Updated instrumentations to use `opentelemetry.trace.use_span` instead of `Tracer.use_span()`
+  ([#364](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/364))
+- `opentelemetry-instrumentation-sqlalchemy` Fix multithreading issues in recording spans from SQLAlchemy ([#315](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/315))
 
 ### Changed
+- Rename `IdsGenerator` to `IdGenerator`
+  ([#350](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/350))
+
+## [0.18b0](https://github.com/open-telemetry/opentelemetry-python-contrib/releases/tag/v0.18b0) - 2021-02-16
+
+### Added
+- `opentelemetry-propagator-ot-trace` Add OT Trace Propagator
+  ([#302](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/302))
+- `opentelemetry-instrumentation-logging` Added logging instrumentation to enable log - trace correlation.
+  ([#345](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/345))
+
+### Removed
 - Remove `component` span attribute in instrumentations. 
   `opentelemetry-instrumentation-aiopg`, `opentelemetry-instrumentation-dbapi` Remove unused `database_type` parameter from `trace_integration` function.
   ([#301](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/301))
-- `opentelemetry-instrumentation-sqlalchemy` Fix multithreading issues in recording spans from SQLAlchemy ([#315](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/315))
+- `opentelemetry-instrumentation-asgi` Return header values using case insensitive keys
+  ([#308](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/308))
+- Remove metrics from all instrumentations
+  ([#312](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/312))
+- `opentelemetry-instrumentation-boto` updated to set span attributes instead of overriding the resource.
+  ([#310](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/310))
+- `opentelemetry-instrumentation-grpc` Fix issue tracking child spans in streaming responses
+  ([#260](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/260))
+- `opentelemetry-instrumentation-grpc` Updated client attributes, added tests, fixed examples, docs
+  ([#269](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/269))
 
 ## [0.17b0](https://github.com/open-telemetry/opentelemetry-python-contrib/releases/tag/v0.17b0) - 2021-01-20